refactor(ui5-datepicker): split tests & do not navigate during execution
