+++ packages/main/test/sap/ui/webcomponents/main/pages/DatePicker_test_page.html
@@ -18,21 +18,21 @@
 </head>
 <body>
 	<ui5-datepicker id="dp"></ui5-datepicker>
+	<ui5-datepicker id="dp10"></ui5-datepicker>
 	<ui5-datepicker id="dp1" value="11/11/11"></ui5-datepicker>
+	<ui5-datepicker id="dp11" value="11/11/11"></ui5-datepicker>
 	<ui5-datepicker id="dp2" format-pattern="yyyy, dd/MM"></ui5-datepicker>
+	<ui5-datepicker id="dp12" format-pattern="yyyy, dd/MM"></ui5-datepicker>
 	<ui5-datepicker id="dp3"></ui5-datepicker>
+	<ui5-datepicker id="dp13"></ui5-datepicker>
+	<ui5-datepicker id="dp14"></ui5-datepicker>
+	<ui5-datepicker id="dp15"></ui5-datepicker>
+	<ui5-datepicker id="dp16"></ui5-datepicker>
 	<ui5-datepicker id="dp4" value="Jan 29, 2019" format-pattern="MMM d, y"></ui5-datepicker>
-	<ui5-datepicker id="dp5"></ui5-datepicker>
 	<ui5-datepicker id="dp6"></ui5-datepicker>
+	<ui5-datepicker id="dp17"></ui5-datepicker>
 	<ui5-datepicker id="dp7" format-pattern="MMM d, y" value="Jan 1, 2019"></ui5-datepicker>
-	<ui5-datepicker id="dp7_1" format-pattern="MMM d, y"></ui5-datepicker>
 	<ui5-datepicker id="dp7_2" format-pattern="MMM d, y"></ui5-datepicker>
-	<ui5-datepicker id="dp8" value="Jan 6, 2015"></ui5-datepicker>
-	<ui5-datepicker id="dp9" value="today"></ui5-datepicker>
-	<ui5-datepicker id="dp10" disabled></ui5-datepicker>
-	<ui5-datepicker id="dp11"></ui5-datepicker>
-	<ui5-datepicker id="dp12"></ui5-datepicker>
-	<label id='lbl'>0</label>
 	<label for="inputTimezone">Timezone:</label>
 	<input id="inputTimezone" type="number"/>
 	<button id="btnApplyTimezone" onclick="stubTimezone();">Apply</button>
@@ -54,14 +54,6 @@
 			Date.prototype.getDate = originalGetDate;
 			document.querySelector("#inputTimezone").value = "";
 		};
-
-		let counter = 0;
-		function increaseCounter() {
-			document.querySelector('#lbl').innerHTML = ++counter;
-		}
-
-		document.querySelector("#dp5").addEventListener("change", increaseCounter);
-		document.querySelector("#dp8").addEventListener("change", increaseCounter);
 	</script>
 </body>
 </html>
