+++ packages/main/test/sap/ui/webcomponents/main/pages/DatePicker_test_page_bg.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<meta http-equiv="X-UA-Compatible" content="IE=edge">
+	<meta charset="utf-8">
+
+	<script data-id="sap-ui-config" type="application/json">
+		{
+			"language": "BG"
+		}
+	</script>
+
+	<script src="../../../../../../resources/sap/ui/webcomponents/main/bundle.esm.js"
+		type="module"
+	>
+	</script>
+</head>
+<body>
+	<ui5-datepicker id="dp" format-pattern="MMM d, y"></ui5-datepicker>
+</body>
+</html>
